requirements.txt
Django==5.1.7
django-admin-sortable2==2.2.4
django-tinymce==4.1.0
pillow==11.1.0
python-environ==0.4.54



# asgiref==3.8.1
# Django==5.1.7
# django-admin-sortable2==2.2.4
# django-tinymce==4.1.0
# environ==1.0
# gunicorn==23.0.0
# packaging==24.2
# pillow==11.1.0
# python-environ==0.4.54
# sqlparse==0.5.3